Add files via upload
Teachers' Day in India is celebrated on September 5th each year to honor the birth anniversary of Dr. Sarvepalli Radhakrishnan, a revered educator and the second President of India. This day acknowledges the invaluable contributions of teachers in shaping the future of students and society. Schools and colleges organize various activities, including speeches, cultural programs, and the presentation of gifts to teachers, expressing gratitude and respect for their hard work and dedication. It serves as a reminder of the importance of education and the pivotal role teachers play in guiding and inspiring the next generation.